Feature/simulations doku

+++ 04_Simulation/ASM/slprj/accel/ASM/rtwtypes.h
@@ -17,11 +17,11 @@ typedef void * pointer_T ;
 #endif
 #ifndef INT64_T
 #define INT64_T
-typedef long long int64_T ;
+typedef long int64_T ;
 #endif
 #ifndef UINT64_T
 #define UINT64_T
-typedef unsigned long long uint64_T ;
+typedef unsigned long uint64_T ;
 #endif
 #ifndef CINT64_T
 #define CINT64_T
